KAA-1233 Fix code review
