+++ client/client-multi/client-c/tools/kaa_encryption/rsa_key_gen/CMakeLists.txt
@@ -18,7 +18,7 @@ cmake_minimum_required(VERSION 2.8.12)
 project(kaa_rsa_keygen C)
 
 add_library(rsa_keygen kaa_rsa_key_gen.c)
-target_include_directories(rsa_keygen PUBLIC ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(rsa_keygen PUBLIC ${CMAKE_CURRENT_LIST_DIR}/include)
 
 add_subdirectory(${KAA_SDK_DIR}/thirdparty/mbedtls kaa_mbedtls)
 target_link_libraries(rsa_keygen kaa_mbedtls)
